feat: Add refactoring tools for incremental rename and general refactoring
This commit adds the refactoring tools for incremental rename and general refactoring to the LazyVim repository. The "smjonas/inc-rename.nvim" plugin is included for incremental rename functionality, which can be triggered with the "<leader>n" keymap. Additionally, the "ThePrimeagen/refactoring.nvim" plugin is added for general refactoring, which can be triggered with the "<leader>rf" keymap in visual mode.

Closes #XYZ